memperbaiki dosen login

+++ app/Http/Controllers/CRUD/DosenLoginController.php
@@ -8,6 +8,7 @@ use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\CrudResource;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Validator;
 
@@ -48,34 +49,48 @@ class DosenLoginController extends Controller
      */
     public function store($request)
     {
-        $data_req = $request;
-        // membuat password
-        $characters = '0123456789abcdefghijklmnopqrstuvwxyz-_#@';
-        $password = '';
-        for ($i = 0; $i < 8; $i++) {
-            $index = rand(0, strlen($characters) - 1);
-            $password .= $characters[$index];
+        DB::beginTransaction();
+        try {
+            $data_req = $request;
+            // membuat password
+            $characters = '0123456789abcdefghijklmnopqrstuvwxyz-_#@';
+            $password = '';
+            for ($i = 0; $i < 8; $i++) {
+                $index = rand(0, strlen($characters) - 1);
+                $password .= $characters[$index];
+            }
+            // membuang karakter setelah , dan .
+            $remove_koma = str_replace(['.', ','], '', $data_req['nama']);
+            $slug = Str::slug($remove_koma, '_'); // Membuat slug dari nama (misalnya: john-doe)
+            $email = Str::finish($slug, '@fstuogp.com'); // Menambahkan "@fstuogp.com" di belakang slug
+            // input data user
+            $user = User::create([
+                'id' => $data_req['id'],
+                'name' => $data_req['nama'],
+                'email' => $email,
+                'password' => Hash::make($password),
+                'show_password' => $password,
+                'role' => 'dosen',
+            ]);
+            // simpan data dosen login
+            DosenLogin::create([
+                'dosen_id' => $data_req['id'],
+                'user_id' => $user->id,
+            ]);
+
+            $data = DosenLogin::latest()->first();
+            DB::commit();
+            return new CrudResource('success', 'Data Berhasil Disimpan', $data);
+        } catch (\Throwable $th) {
+            // jika terdapat kesalahan
+            DB::rollback();
+            $message = [
+                'judul' => 'Gagal',
+                'type' => 'error',
+                'message' => $th->getMessage(),
+            ];
+            return response()->json($message, 400);
         }
-        $slug = Str::slug($data_req['nama'], '_'); // Membuat slug dari nama (misalnya: john-doe)
-        $email = Str::finish($slug, '@fstuogp.com'); // Menambahkan "@fstuogp.com" di belakang slug
-        // input data user
-        $user = User::create([
-            'id' => $data_req['id'],
-            'name' => $data_req['nama'],
-            'email' => $email,
-            'password' => Hash::make($password),
-            'show_password' => $password,
-            'role' => 'dosen',
-        ]);
-        // simpan data dosen login
-        DosenLogin::create([
-            'dosen_id' => $data_req['id'],
-            'user_id' => $user->id,
-        ]);
-
-        $data = DosenLogin::latest()->first();
-
-        return new CrudResource('success', 'Data Berhasil Disimpan', $data);
     }
 
     /**
@@ -126,6 +141,10 @@ class DosenLoginController extends Controller
     public function destroy($id)
     {
         $data = DosenLogin::where('user_id', $id)->first();
+        // jika data dosen login tidak ada
+        if (!$data) {
+            return new CrudResource('error', 'Data Tidak Ditemukan', $data);
+        }
         // delete data user
         User::destroy($id);
         // delete data
